finder: fix for panic on empty seriesByTag() call
